# config: Relax limitation of tso-update-physical-interval