Issue Number: #5440

Signed-off-by: pingyu <yuping@pingcap.com>

<!--
Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed
-->

### What problem does this PR solve?

<!--

Please create an issue first to describe the problem.
There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#linking-issues.

-->
Issue Number: Close #5440

### What is changed and how does it work?

<!--

You could use the "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/pd/blob/master/CONTRIBUTING.md#format-of-the-commit-message.

-->

```commit-message
Minimal valid value of config item `tso-update-physical-interval` is changed to 1ms.
```

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of these tests must be included. -->

- Unit test
- Integration test
- Benchmark

| tso-update-physical-interval | CPU without workload | CPU with pd-tso-bench | pd-tso-bench QPS | pd-tso-bench P99 latency |
|------------------------------|----------------------|-----------------------|------------------|--------------------------|
| 1 ms                         | 15%                  | 28%                   | 458306           | 3.59 ms                  |
| 3 ms                         | 9.3%                 | 25%                   | 470455           | 3.57 ms                  |
| 5 ms                         | 7.6%                 | 25%                   | 354989           | 8.64 ms                  |
| 50 ms                        | 6%                   | 7%                    | 44466            | 54.1 ms                  |

Notes:
1. The benchmark is performed on a 40C 125GB NVMe machine, with 1 pd, 4 kv, 1 db, deployed by TiUP playground.
2. `count` in TSO request is set to 100x (`uint32(count) * 100` [here](https://github.com/tikv/pd/blob/6d883d0e8386c64ac12ade25d00893a608b33a14/client/client.go#L1072)) to make it overflow the logical interval.
3. There is no significant difference on memory usage.
4. The benchmark showed that `3ms` is good enough, and `1ms` has no improvement but just causes more CPU. But `1ms` can still provide 3 times TSOs than `3ms`, so `1ms` is still a valid value.

Code changes

- Has configuration change
`tso-update-physical-interval`

Side effects

Related changes

- PR to update [`pingcap/docs`](https://github.com/pingcap/docs)/[`pingcap/docs-cn`](https://github.com/pingcap/docs-cn): TBD
- PR to update [`pingcap/tiup`](https://github.com/pingcap/tiup): TBD


### Release note

<!-- A bugfix or a new feature needs a release note. If there is no need release note, just uncomment the below line. -->

```release-note
Minimal valid value of config item `tso-update-physical-interval` is changed to 1ms.
```